requirements.txt
# Core dependencies (updated for Python 3.13 compatibility)
langchain>=0.1.0
langchain-google-genai>=0.0.6
langchain-anthropic>=0.1.1
pydantic>=2.10.0

# Google APIs
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pytz==2023.3

# Optional: Web interface
gradio==4.13.0

# Optional: Fireflies integration
requests==2.31.0

# MCP and Whisper for transcription
mcp>=1.0.0
openai-whisper>=20231117
